+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 * keep polling for 10 Minutes after engine stop to solve: detect car is charging #20
+* battery target load
 TODO: work on (remoteClimateState)[https://github.com/orgs/SAIC-iSmart-API/discussions/6#discussioncomment-5923522] related to  Feature request: A/C control #13
 
 ## 0.2.4